[Qt] QFlag won't take a ulong on an LP64 environment. Contributes JB#36790

+++ src/extensions/qcontactstatusflags_impl.h
@@ -55,7 +55,7 @@ void QContactStatusFlags::setFlags(Flags flags)
 
 QContactStatusFlags::Flags QContactStatusFlags::flags() const
 {
-    return Flags(static_cast<ulong>(flagsValue()));
+    return Flags(static_cast<uint>(flagsValue()));
 }
 
 void QContactStatusFlags::setFlagsValue(quint64 value)
